+++ deepparse/data_error/data_error.py
@@ -4,6 +4,7 @@ class DataError(Exception):
     """
 
     def __init__(self, value):
+        super().__init__()
         self.value = value
 
     def __str__(self):