+++ deepparse/parser/address_parser.py
@@ -23,10 +23,9 @@ from .tools import (
     indices_splitting,
     handle_model_name,
 )
-from .. import validate_data_to_parse, whitespace_only_addresses
+from .. import validate_data_to_parse
 from ..converter import TagsConverter
 from ..converter import fasttext_data_padding, bpemb_data_padding, DataTransform
-from ..data_error.data_error import DataError
 from ..dataset_container import DatasetContainer
 from ..embeddings_models import BPEmbEmbeddingsModel
 from ..embeddings_models import FastTextEmbeddingsModel
@@ -244,6 +243,7 @@ class AddressParser:
             addresses_to_parse (Union[list[str], str]): The addresses to be parsed, can be either a
                 single address (when using str) or a list of address. Validation tests on the dataset to parse are
                 done in order to validate the following basic criteria:
+
                     - no address are empty string, and
                     - no address are whitespace only string.
 
@@ -350,7 +350,10 @@ class AddressParser:
         prediction tags, and if new ``seq2seq_params`` were used, the new seq2seq parameters.
 
         Args:
-            dataset_container (~deepparse.dataset_container.DatasetContainer): The dataset container of the data to use.
+            dataset_container (~deepparse.dataset_container.DatasetContainer): The dataset container of the data to use
+                such as any PyTorch Dataset (:class:`~torch.utils.data.Dataset`) user define class or one of our two
+                DatasetContainer (:class:`~deepparse.dataset_container.PickleDatasetContainer` or
+                :class:`~deepparse.dataset_container.CSVDatasetContainer`)
             train_ratio (float): The ratio to use of the dataset for the training. The rest of the data is used for the
                 validation (e.g. a train ratio of 0.8 mean a 80-20 train-valid split) (default is 0.8).
             batch_size (int): The size of the batch (default is 32).
@@ -419,9 +422,9 @@ class AddressParser:
                 import poutyne
 
                 address_parser = AddressParser(device=0)
-                data_path = 'path_to_a_pickle_dataset.p'
+                data_path = 'path_to_a_csv_dataset.p'
 
-                container = PickleDatasetContainer(data_path)
+                container = CSVDatasetContainer(data_path)
 
                 lr_scheduler = poutyne.StepLR(step_size=1, gamma=0.1) # reduce LR by a factor of 10 each epoch
                 address_parser.retrain(container, 0.8, epochs=5, batch_size=128, callbacks=[lr_scheduler])